edit google sign out functionality

+++ app/src/main/java/todolistmvp/modul/home/HomePresenter.java
@@ -1,8 +1,15 @@
 package todolistmvp.modul.home;
 
 
+import android.util.Log;
+import android.widget.Toast;
+
+import androidx.annotation.Nullable;
+
 import com.google.android.gms.auth.api.Auth;
 import com.google.android.gms.common.api.GoogleApiClient;
+import com.google.android.gms.common.api.ResultCallback;
+import com.google.android.gms.common.api.Status;
 import com.google.firebase.auth.FirebaseAuth;
 
 import java.util.ArrayList;
@@ -16,12 +23,12 @@ public class HomePresenter implements HomeContract.Presenter{
     private final HomeContract.View view;
     private final SessionRepository sessionRepository;
     private final TableHandler tableHandler;
-    private GoogleApiClient googleApiClient;
 
     public HomePresenter(HomeContract.View view, SessionRepository sessionRepository, TableHandler tableHandler) {
         this.view = view;
         this.sessionRepository = sessionRepository;
         this.tableHandler = tableHandler;
+
     }
 
     @Override
@@ -50,19 +57,17 @@ public class HomePresenter implements HomeContract.Presenter{
     }
 
     @Override
-    public void performLogout() {
+    public void performLogout(@Nullable GoogleApiClient googleApiClient) {
         FirebaseAuth.getInstance().signOut();
-//        Auth.GoogleSignInApi.signOut(googleApiClient).setResultCallback(
-//                new ResultCallback<Status>() {
-//                    @Override
-//                    public void onResult(Status status) {
-//                        if (status.isSuccess()){
-//                            gotoMainActivity();
-//                        }else{
-//                            Toast.makeText(getApplicationContext(),"Session not close",Toast.LENGTH_LONG).show();
-//                        }
-//                    }
-//                });
-        sessionRepository.destroy();
+        Auth.GoogleSignInApi.signOut(googleApiClient).setResultCallback(
+            new ResultCallback<Status>() {
+                public void onResult(Status status) {
+                    if (status.isSuccess()){
+                        sessionRepository.destroy();
+                    }else{
+                        Log.d("error: ", "Session not closed");
+                    }
+                }
+            });
     }
 }
